외부 요청을 처리하는 객체 인터페이스로 구현
