--- src/main/java/uhs/alphabet/domain/badge/codeforces/BadgeStream.java
@@ -1,5 +1,0 @@
-package uhs.alphabet.domain.badge.codeforces;
-
-public interface BadgeStream {
-    public String getBadge();
-}
